Added a GameMode Move property and 2 functions
* Fixed an issue with GameMode Items in battles
* Updated the example files for GameMode Items and GameMode Moves
* Maybe fixed something audio volume related

* Made it possible to disable the reduction of PP using the DeductPP (bool) property for GameModeMoves
* Added the new functions "SetTrigger" and "RemoveTrigger" for GameModeMoves:

Function "SetTrigger" checks if a Pokémon does or does not have a certain Status and prevents functions after it from executing if the Pokémon does not have the desired Status.
The function is formatted like this: Function|SetTrigger,Target,Triggers
* "Target" determines which Pokémon's Status is checked and has to be either 0 or 1 (0 = check own Pokémon, 1 = check opponent's Pokémon)
* "Triggers" determines the status effect(s) that should or should not be on the target Pokémon. You can add multiple possible status effects to check for by separating them with a semicolon (;), the values can be: "burn", "freeze", "paralyze", "poison", "badpoison", "anypoison", "sleep", "noburn", "nofreeze", "noparalyze", "nopoison", "nobadpoison", "nopoison", "nosleep"
Function "RemoveTrigger" removes such a trigger so that functions after it are executed

+++ P3D/Core/GameOptions.vb
@@ -37,12 +37,12 @@
 
                 Select Case name.ToLower()
                     Case "volume"
-                        MusicManager.MasterVolume = CSng(CDbl(value) / 100)
-                        SoundManager.Volume = CSng(CDbl(value) / 100)
+                        MusicManager.MasterVolume = CSng(CInt(value) / 100)
+                        SoundManager.Volume = CSng(CInt(value) / 100)
                     Case "music"
-                        MusicManager.MasterVolume = CSng(CDbl(value) / 100)
+                        MusicManager.MasterVolume = CSng(CInt(value) / 100)
                     Case "sound"
-                        SoundManager.Volume = CSng(CDbl(value) / 100)
+                        SoundManager.Volume = CSng(CInt(value) / 100)
                     Case "muted"
                         SoundManager.Muted = CBool(value)
                         MusicManager.Muted = CBool(value)
